+++ server/voice_changer/RVC/inferencer/WebUIInferencer.py
@@ -50,5 +50,5 @@ class WebUIInferencer(Inferencer):
             return_length=return_length,
             formant_length=formant_length
         )
-        res = res[0][0, 0].float()
+        res = res[0][0, 0]
         return torch.clip(res, -1.0, 1.0)
